Update SendEmail, Validation

+++ QLPhongMach/src/main/webapp/resources/js/thuocs.js
@@ -3,22 +3,25 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/JavaScript.js to edit this template
  */
 function deleteThuoc(endpoint, id, btn) {
-    let r = document.getElementById(`row${id}`);
-    let load = document.getElementById(`load${id}`);
-    load.style.display = "block";
-    btn.style.display = "none";
-    fetch(endpoint, {
-        method: 'delete'
-    }).then(function (res) {
-        if (res.status !== 204)
-            alert("Something wrong!!!");
-        load.style.display = "none";
-        r.style.display = "none";
-    }).catch(function (err) {
-        console.error(err);
-        btn.style.display = "block";
-        load.style.display = "none";
-    });
+    if (confirm("Ban chac chan xoa thuoc nay khong?") === true) {
+        let r = document.getElementById(`row${id}`);
+        let load = document.getElementById(`load${id}`);
+        load.style.display = "block";
+        btn.style.display = "none";
+        fetch(endpoint, {
+            method: 'delete'
+        }).then(function (res) {
+            if (res.status !== 204)
+                alert("Something wrong!!!");
+            load.style.display = "none";
+            r.style.display = "none";
+        }).catch(function (err) {
+            console.error(err);
+            btn.style.display = "block";
+            load.style.display = "none";
+        });
+    }
+    
 }
 
 function getThuocs(endpoint) {
@@ -37,7 +40,7 @@ function getThuocs(endpoint) {
                     <td>${data[i].ghiChu}</td>
                     <td>
                         <div class="spinner-border text-info" style="display:none" id="load${data[i].id}"></div>
-                        <button class="btn btn-danger" onclick="deleteThuoc('${endpoint + "/" + data[i].id}', ${data[i].id}, this)">Xoa</button>
+                        <button class="btn btn-danger" onclick="deleteThuoc('${endpoint + "/" + data[i].id}', ${data[i].id}, this)">Xoá</button>
                     </td>
                 </tr>
             `
